requirements.txt
openai==1.3.0
flask==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
httpx==0.28.1
